Scale spatial params to resolution across ops; fix delete/backspace node deletion
Apply scale_to_resolution to spatial size/radius params across ~60 image
operations (blur, morphology, stylize, smoothing, fx, transform, etc.) so
authored values produce the same relative effect at any resolution, and
update input descriptions to note the 1024px reference.

Also accept Backspace (not just Delete) to delete selected nodes, since on
macOS the key labelled "delete" is Backspace; guarded so it doesn't fire
while a text field has keyboard focus.

+++ app/crates/mangler_core/src/operations/images/filter/morphology/outline_tests.rs
@@ -145,7 +145,8 @@ async fn strokes_alpha_defined_shape_with_bright_rgb() {
     // no edge here and emitted a fully-transparent (empty) image.
     let mut inputs = vec![
         Input::new("image".into(), Value::Image { data: alpha_square(32, 32), change_id: get_id() }, None, None),
-        Input::new("thickness".into(), Value::Integer(2), None, None),
+        // thickness is reference-px (at 1024); 64 → 2px effective on this 32x32 image.
+        Input::new("thickness".into(), Value::Integer(64), None, None),
         Input::new("position".into(), Value::Integer(2), None, None),
         Input::new("color".into(), black(), None, None),
     ];
@@ -171,7 +172,8 @@ async fn outer_ring_thickness_is_uniform_around_a_circle() {
     let size = 96;
     let mut inputs = vec![
         Input::new("image".into(), Value::Image { data: filled_circle(size, 24.0), change_id: get_id() }, None, None),
-        Input::new("thickness".into(), Value::Integer(8), None, None),
+        // thickness is reference-px (at 1024); 85 → ~8px effective on this 96x96 image.
+        Input::new("thickness".into(), Value::Integer(85), None, None),
         Input::new("position".into(), Value::Integer(0), None, None),
         Input::new("color".into(), black(), None, None),
     ];
